refactor(terms): replace the mwc terms with Kabam
Since we are moving along from MWC to kabam, so we need to rename mwc terms on this kernel to avoid confusion. This is necessary since someone else will look into our source code especially the newcomer.
The consequence of this changes is we also need to replace the mwc terms in our plugins source code.